+++ modules/ult/list_horn.lua
@@ -23,7 +23,7 @@ local svDefault = {
     windowScale = 1.0,
     windowPosLeft = 10,
     windowPosTop = 570,
-    windowWidth = 262,
+    windowWidth = 272,
     backgroundOpacity = 0.0,
 
     showPercentValue = 1.0,